requirements.txt
# Python Dependencies for Reddit Newsletter System

# RSS Feed Parsing
feedparser==6.0.11

# LLM APIs (choose one or more)
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Environment Variables
python-dotenv==1.0.0

# Database (built-in to Python 3.x)
# sqlite3 - no install needed

# Utilities
requests>=2.31.0
google-genai>=1.52.0
Pillow
groq

# Resilience
tenacity>=8.2.0

# Style Enforcement
lexicalrichness>=0.5
nltk>=3.8

# Web UI
streamlit>=1.30.0
